fix: moved product and category models out of the enums
