+++ src/main/java/com/amalitech/gpuconfigurator/repository/CategoryRepository.java
@@ -1,5 +1,6 @@
 package com.amalitech.gpuconfigurator.repository;
 
+import com.amalitech.gpuconfigurator.model.Category;
 import com.amalitech.gpuconfigurator.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,12 +11,12 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface CategoryRepository extends JpaRepository<Role.Category, UUID> {
+public interface CategoryRepository extends JpaRepository<Category, UUID> {
 
     @Query("SELECT s FROM Category s WHERE s.categoryName = ?1 ")
-   Optional<Role.Category> findByCategoryName(String categoryName);
+   Optional<Category> findByCategoryName(String categoryName);
 
     @Query("SELECT s FROM Category s WHERE s.categoryName = ?1 ")
-    List<Role.Category> findByCategoryNameList(String categoryName);
+    List<Category> findByCategoryNameList(String categoryName);
 
 }
