+++ src/main/java/com/amalitech/gpuconfigurator/repository/ProductRepository.java
@@ -1,6 +1,7 @@
 package com.amalitech.gpuconfigurator.repository;
 
 import com.amalitech.gpuconfigurator.model.Otp;
+import com.amalitech.gpuconfigurator.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
@@ -10,12 +11,12 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Otp.Product, UUID> {
+public interface ProductRepository extends JpaRepository<Product, UUID> {
 
 
     @Modifying
     void deleteById(UUID id);
 
     @Query("SELECT i FROM Product i WHERE i.productId =?1")
-    Optional<Otp.Product> getProductByProductId(String productId);
+    Optional<Product> getProductByProductId(String productId);
 }
